backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
bcrypt==4.1.3
PyJWT==2.8.0
pandas==2.2.2
scikit-learn==1.4.2
numpy==1.26.4
